+++ src/error.rs
@@ -1,12 +1,12 @@
 use crate::launch::ProcfileConversionError;
-use crate::procfile::ProcfileParsingError;
+use crate::procfile::ProcfileError;
 use bullet_stream::Print;
 use indoc::formatdoc;
 
 #[derive(Debug)]
 pub(crate) enum ProcfileBuildpackError {
     CannotReadProcfileContents(std::io::Error),
-    ProcfileParsingError(ProcfileParsingError),
+    ProcfileParsingError(ProcfileError),
     ProcfileConversionError(ProcfileConversionError),
 }
 
@@ -23,9 +23,16 @@ pub(crate) fn error_handler(buildpack_error: ProcfileBuildpackError) {
                 Underlying cause was: {io_error}
             "});
         }
-        // There are currently no ways in which parsing can fail, however we will add some in the future:
-        // https://github.com/heroku/buildpacks-procfile/issues/73
-        ProcfileBuildpackError::ProcfileParsingError(parsing_error) => match parsing_error {},
+        ProcfileBuildpackError::ProcfileParsingError(parsing_error) =>
+            build_output.error(formatdoc! {"
+                Invalid Procfile format
+
+                The provided `Procfile` contains an invalid format and the buildpack cannot continue.
+
+                To fix this problem please correct the following error and commit the results to git:
+
+                {parsing_error}
+            "}),
         ProcfileBuildpackError::ProcfileConversionError(conversion_error) => match conversion_error
         {
             ProcfileConversionError::InvalidProcessType(libcnb_error) => {